* Added Citizens of the Imperium careers Added cascadeAircraft, cascadeBow, and cascadeWatercraft. Added cascadeServiceAircraft to handle the Flyer's initial Aircraft service skill Added draftServices and make the six options for draft Army, Flyers, Marines, Navy, Sailors, and Scouts.
For random service generation, if initial social level is 10 or better there is a 1 in 3 chance of the character opting for the Noble career.

If a character is already social standing 11 (Knight) or 12 (Baron) they are automatically promoted to those ranks in the noble career. I kept Commissioned false so there a still a possibility of a two rank bump in the first term.

Belters, Doctors, Rogues, Scientists, and Hunters get two skill per term.